test(#879): consolidate the PostHog env scrub into the root tests conftest
The med review round flagged the near-verbatim duplication of the session-scoped scrub_posthog_env fixture across the integration and e2e conftests as a drift hazard — a future change to the scrub landing in one tier but not the other would silently leave the unpatched tier able to send real discogs_rate_gate_fail_open events from a developer machine. tests/conftest.py is already the shared cross-tier fixture home, so the two copies collapse into one session fixture there; the unit tier's own scrub_credential_env already blanks the key, for which the extra blank is a no-op.
